changes with new robot

+++ pybricks/pyBricks1/shared.py
@@ -5,9 +5,10 @@ from pybricks.robotics import DriveBase
 
 class Speed:
     Straight = None
-    Slow = float(100)
-    Medium = float(200)
-    Fast = float(400)
+    Slow = float(150)
+    Medium = float(300)
+    # 600 is max, but sometimes gyro make mistakes so, I guess 500 makes little mistake so 450 is sweet spot
+    Fast = float(450)
 
 
 # inspiration https://github.com/FLL-Team-24277/FLL-Fall-2023-Masterpiece/blob/main/samples/base_robot.py
